Build error due to double Info.plist
We use CocoaPods for dependency management and very often get the error "Failed to verify code signature of /path/to/on/physical/device/MyApp.app/Frameworks/SlideMenuControllerSwift.framework/SlideMenuControllerSwift" or "required code signature missing for ..." when deploying the app to a physical device.
According to this (http://stackoverflow.com/questions/40110392/required-code-signature-missing-for-a-library) we are not the only one.
After finding this (https://github.com/vikmeup/SCLAlertView-Swift/issues/114) I assume it has to do with the source_files pattern in your podspec, leading to a duplicate Info.plist. This results in undefined behavior on which Info.plist is included in the final framework bundle. Sometimes the build system selects the wrong one which isn't preprocessed and still contains all the placeholders like ${EXECUTABLE_NAME}.